resolve session error, resolve duplicate ids
