+++ app/builder/loading.tsx
@@ -1,0 +1,9 @@
+export default function Loading() {
+  return (
+    <div className="flex items-center justify-center min-h-screen">
+      <div className="text-lg font-medium text-gray-600">
+        Loading resume builder...
+      </div>
+    </div>
+  )
+}
