+++ middleware.ts
@@ -1,0 +1,11 @@
+import { withAuth } from "next-auth/middleware"
+
+export default withAuth({
+  callbacks: {
+    authorized: ({ token }) => !!token
+  },
+})
+
+export const config = {
+  matcher: ["/builder"]
+}
