+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@auth/core": "^0.34.2",
     "@radix-ui/react-label": "^2.1.1",
     "@radix-ui/react-slot": "^1.1.1",
     "@types/lodash": "^4.17.14",
@@ -18,8 +19,6 @@
     "lodash": "^4.17.21",
     "lucide-react": "^0.471.1",
     "mongoose": "^8.9.5",
-    "next": "^15.1.4",
-    "next-auth": "^4.24.11",
     "next-transpile-modules": "^10.0.1",
     "react": "^19.0.0",
     "react-dnd": "^16.0.1",
@@ -29,13 +28,15 @@
     "react-merge-refs": "^2.1.1",
     "remark-gfm": "^4.0.0",
     "tailwind-merge": "^2.6.0",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "uuid": "^11.0.5"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
+    "@types/uuid": "^10.0.0",
     "autoprefixer": "^10.4.20",
     "eslint": "^9",
     "eslint-config-next": "15.1.4",
